change action to submit
