--- github/helpers/githubActivityGraphURL.ts
@@ -1,4 +1,0 @@
-
-export function githubActivityGraphUrl(username : string): string {
-    return `https://activity-graph.herokuapp.com/graph?username=${username}&bg_color=ffffff&color=708090&line=24292e&point=24292e`
-}
